Remove java prefix from package name
