+++ LabApiUtilities/src/main/com/microsoft/identity/labapi/utilities/authentication/IConfidentialAuthClient.java
@@ -20,10 +20,10 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package java.com.microsoft.identity.labapi.utilities.authentication;
+package com.microsoft.identity.labapi.utilities.authentication;
 
-import java.com.microsoft.identity.labapi.utilities.authentication.common.port.CertificateCredential;
-import java.com.microsoft.identity.labapi.utilities.authentication.common.port.ClientAssertion;
+import com.microsoft.identity.labapi.utilities.authentication.port.CertificateCredential;
+import com.microsoft.identity.labapi.utilities.authentication.port.ClientAssertion;
 
 public interface IConfidentialAuthClient {
